backend/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
gunicorn>=20.0
Flask-Bcrypt>=0.7
Flask-JWT-Extended>=4.2
Flask-cors
Flask-Swagger-UI
PyYAML